requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
matplotlib==3.8.2
statsmodels==0.14.1
pyyaml==6.0.1

# Explainability
shap==0.44.1

# Google Trends
pytrends==4.9.2

# BERT Sentiment
transformers==4.36.2
torch==2.1.2
prophet==1.1.5
streamlit==1.26.1